Rename Arbor#root to Arbor#state
This is so we don't expose internal data structure concepts to the outside world. The fact that Arbor uses a state tree internally does not mean users of the lib need to know that. Users are usually more familiar with concepts like 'store', 'state', thus the motivation for this change