Ensure node values are seeded at very specific points
There are two points in the OST creation where a node needs to be seeded. 1) When traversing from a parent node to its child for the first time, where the child node needs to be created and cached. 2) when setting the root state of the store where the Path is empty. With that said, I think we may be able to change how we handle paths so that even the root node's seed is included in the path.